# math: Pow(x,y) optimization for cases: constant integer `y`. 