Fixes #29456