@Konstantin8105 2018-12-30
Benchstat:
```
name                   old time/op  new time/op  delta
/math.Pow(+128.000)-4  68.7ns ± 0%  67.3ns ± 0%   -2.04%  (p=0.000 n=10+6)
/math.Pow(+127.000)-4  60.0ns ± 0%  49.8ns ± 1%  -17.04%  (p=0.000 n=9+8)
/math.Pow(+126.000)-4  60.8ns ± 0%  47.9ns ± 0%  -21.20%  (p=0.000 n=10+8)
/math.Pow(+124.000)-4  61.4ns ± 0%  45.5ns ± 0%  -26.02%  (p=0.000 n=10+8)
/math.Pow(+100.000)-4  62.9ns ± 1%  35.7ns ± 0%  -43.24%  (p=0.000 n=9+7)
/math.Pow(+99.000)-4   61.8ns ± 0%  34.8ns ± 5%  -43.74%  (p=0.000 n=10+10)
/math.Pow(+19.000)-4   57.0ns ± 0%  25.8ns ± 2%  -54.72%  (p=0.000 n=10+9)
/math.Pow(+10.000)-4   54.9ns ± 0%  23.6ns ± 1%  -56.91%  (p=0.000 n=9+8)
/math.Pow(+5.000)-4    51.9ns ± 0%  21.2ns ± 0%  -59.04%  (p=0.000 n=10+9)
/math.Pow(+2.000)-4    50.1ns ± 0%  18.9ns ± 0%  -62.28%  (p=0.008 n=6+7)
/math.Pow(-2.000)-4    47.4ns ± 0%  16.8ns ± 0%  -64.47%  (p=0.000 n=10+10)
/math.Pow(-5.000)-4    49.1ns ± 0%  18.9ns ± 1%  -61.49%  (p=0.000 n=9+9)
/math.Pow(-10.000)-4   52.3ns ± 0%  21.6ns ± 0%  -58.66%  (p=0.000 n=10+8)
/math.Pow(-19.000)-4   54.2ns ± 0%  23.3ns ± 3%  -57.13%  (p=0.000 n=10+10)
/math.Pow(-99.000)-4   59.0ns ± 0%  33.2ns ± 1%  -43.74%  (p=0.000 n=9+8)
/math.Pow(-100.000)-4  60.2ns ± 0%  32.7ns ± 5%  -45.66%  (p=0.000 n=10+10)
/math.Pow(-124.000)-4  59.1ns ± 1%  43.6ns ± 0%  -26.29%  (p=0.000 n=10+8)
/math.Pow(-126.000)-4  59.0ns ± 0%  44.6ns ± 4%  -24.47%  (p=0.000 n=10+10)
/math.Pow(-127.000)-4  59.0ns ± 0%  46.5ns ± 1%  -21.19%  (p=0.000 n=10+9)
/math.Pow(-128.000)-4  65.9ns ± 0%  23.9ns ± 4%  -63.71%  (p=0.000 n=10+10)
/math.Pow(+128.100)-4   144ns ± 0%   143ns ± 1%   -0.83%  (p=0.000 n=7+10)
/math.Pow(+127.100)-4   139ns ± 0%   139ns ± 0%     ~     (p=0.137 n=8+10)
/math.Pow(+126.100)-4   139ns ± 0%   139ns ± 0%     ~     (all equal)
/math.Pow(+124.100)-4   139ns ± 0%   139ns ± 0%     ~     (all equal)
/math.Pow(+100.100)-4   139ns ± 1%   139ns ± 0%     ~     (p=0.173 n=10+9)
/math.Pow(+99.100)-4    139ns ± 0%   139ns ± 0%     ~     (all equal)
/math.Pow(+19.100)-4    134ns ± 1%   134ns ± 0%     ~     (p=0.337 n=10+10)
/math.Pow(+10.100)-4    130ns ± 1%   128ns ± 0%   -1.77%  (p=0.000 n=10+9)
/math.Pow(+5.100)-4     126ns ± 0%   128ns ± 1%   +1.35%  (p=0.000 n=9+10)
/math.Pow(+2.100)-4     123ns ± 0%   125ns ± 0%   +1.63%  (p=0.000 n=8+10)
/math.Pow(-1.900)-4     128ns ± 0%   130ns ± 0%   +1.95%  (p=0.000 n=8+10)
/math.Pow(-4.900)-4     128ns ± 1%   132ns ± 1%   +3.12%  (p=0.000 n=10+10)
/math.Pow(-9.900)-4     129ns ± 1%   132ns ± 1%   +2.32%  (p=0.000 n=10+10)
/math.Pow(-18.900)-4    133ns ± 0%   134ns ± 0%   +0.75%  (p=0.000 n=9+8)
/math.Pow(-98.900)-4    138ns ± 0%   139ns ± 1%   +0.87%  (p=0.000 n=9+10)
/math.Pow(-99.900)-4    138ns ± 1%   139ns ± 0%   +0.87%  (p=0.000 n=10+10)
/math.Pow(-123.900)-4   138ns ± 0%   142ns ± 1%   +2.54%  (p=0.000 n=8+10)
/math.Pow(-125.900)-4   138ns ± 0%   141ns ± 1%   +1.73%  (p=0.000 n=10+10)
/math.Pow(-126.900)-4   139ns ± 0%   141ns ± 0%   +1.68%  (p=0.012 n=8+3)
```