@pjebs 2019-01-11
Why'd the `case y == 1:` case get removed. It's still an optimization.