@Konstantin8105 2019-01-11
Dear @pjebs ,
The approach of that PR, optimization for integer value y = ~127..127. So, for case `y == 1` it is also work.