@pjebs 2019-01-11
I understand that, but having case `y==1` is still _technically_ faster (even if negligible)