@pjebs 2019-01-11
How can you beat this?:
```
case y == 1:	
		return x
```