@Konstantin8105 2019-01-11
Dear @pjebs ,
Yes, you are right.
On one hand, that PR slow down for `case y==1`(but I don't that in practice).
On another hand, that PR better for cases y == 2 ,y == 3 ... (I use it in practice).
So, prepare that PR in according to my practice of using function `math.Pow`.
Also, it is not a **best** way, the best way is `inline function with constant integer y`. For example I create a package for that https://github.com/Konstantin8105/pow with much more performance.
```
name          time/op
/math.Pow2-4   357ns ± 0%
/pow.E2-4     5.69ns ± 0%

/math.Pow3-4   359ns ± 1%
/pow.E3-4     8.86ns ± 1%
```