@pjebs 2019-01-11
You can keep `y == 1` case and your new case.