@Konstantin8105 2019-01-11
Dear @pjebs ,
Thank you for your point of view.
I will not return specific `case y == 1`, because I don't see the reason in practice point of view.